requirement.txt
requests == 2.21.0
elasticsearch == 6.3.1
PyYAML == 3.13